merge dev_supabase (#53)
* updated the calling conditions of verification of ownership for pin
it was causing an error, preventing registration
added a policy to actually post and a trigger to make the user it’s other
added automatic removal of authorless posts

* didn’t i just fix that ????
thought i did. didn’t. properly fixed issue that was « fixed » on last commit, an ill configured trigger that prevented signups.